Move and rename hlk_sw16 base entity to separate module (#126096)
